Introduce static ImageRenderer
- This class contains all static methods from the RenderEngine that are used to draw images on a Graphics2D object.
