+++ src/de/gurkenlabs/litiengine/gui/SpeechBubble.java
@@ -24,7 +24,7 @@ import de.gurkenlabs.litiengine.IUpdateable;
 import de.gurkenlabs.litiengine.entities.IEntity;
 import de.gurkenlabs.litiengine.entities.IEntityProvider;
 import de.gurkenlabs.litiengine.graphics.IRenderable;
-import de.gurkenlabs.litiengine.graphics.RenderEngine;
+import de.gurkenlabs.litiengine.graphics.ImageRenderer;
 import de.gurkenlabs.litiengine.graphics.RenderType;
 import de.gurkenlabs.litiengine.sound.Sound;
 import de.gurkenlabs.litiengine.util.ImageProcessing;
@@ -138,7 +138,7 @@ public class SpeechBubble implements IUpdateable, IRenderable, IEntityProvider {
 
     final float startX = (float) (entityCenter.getX() - deltaX);
     final float startY = (float) (entityCenter.getY() - deltaY);
-    RenderEngine.renderImage(g, this.bubble, new Point2D.Double(startX, startY));
+    ImageRenderer.renderImage(g, this.bubble, new Point2D.Double(startX, startY));
 
     g.setColor(this.getAppearance().getForeColor());
     final FontRenderContext frc = g.getFontRenderContext();
