+++ tests/de/gurkenlabs/litiengine/graphics/RenderEngineTests.java
@@ -30,7 +30,7 @@ public class RenderEngineTests {
     Graphics2D graphics = mock(Graphics2D.class);
 
     final Image img = ImageProcessing.getCompatibleImage(5, 5);
-    RenderEngine.renderImage(graphics, img, new Point2D.Double(10, 20));
+    ImageRenderer.renderImage(graphics, img, new Point2D.Double(10, 20));
 
     ArgumentCaptor<Image> captor = ArgumentCaptor.forClass(Image.class);
     ArgumentCaptor<AffineTransform> transCaptor = ArgumentCaptor.forClass(AffineTransform.class);
